renames files for clarity, adds comments
